html/semantics/scripting-1/the-script-element/scriptexecute-events/001.html

<!DOCTYPE html>
<html>
  <title>beforescriptexecute inline script</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    target = null;
    type = null;
    isTrusted = null;
    document.addEventListener("beforescriptexecute", function(e) {target = e.target; type=e.type; isTrusted=e.isTrusted}, false);
  </script>
  <script>
    test(function() {
           assert_equals(type, "beforescriptexecute");
           assert_equals(target, document.getElementsByTagName("script")[3])
           assert_equals(isTrusted, true);
    });
 </script>
